# laravel_project/laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8089/api",
            "description": "NGNIX serwer"
        }
    ],
    "paths": {
        "/decryptMessage/{message}": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Decrypt message.",
                "operationId": "44e9a51513af3e745cb38457585b5abe",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "Message to decrypt.",
                        "schema": {
                            "type": "String",
                            "example": "(o>!a!"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/encryptMessage/{message}": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Encrypt message.",
                "operationId": "1690daddbf033534d223a42b5c640184",
                "parameters": [
                    {
                        "name": "encryptedMessage",
                        "in": "path",
                        "description": "Message to encrypt.",
                        "schema": {
                            "type": "String",
                            "example": "dziala"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/lcdDisplay/{digits}": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Digits",
                "operationId": "05d712d5575c07f873ed562c9f6134ec",
                "parameters": [
                    {
                        "name": "digits",
                        "in": "path",
                        "description": "Digits to display",
                        "schema": {
                            "type": "String",
                            "example": "0123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/lotteryWinners": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Wygrane losy",
                "operationId": "1eb2dab358cded93aa85042340965d33",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/incomeStatistic": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Statystyka zarobków",
                "operationId": "28efa18632709d4499b553cfb4a35f29",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}